refactor(infra): extract data export/import into InfraDataService with entity-derived table registry
Move the export/import machinery out of InfraDataController into a new
InfraDataService; the controller keeps only routing, role guards, the
validation DTO and the published response metadata. All observable
contracts are unchanged (response shapes, counts keys, skippedTables
semantics, 400/409 codes and their code strings).

The export table set now lives in one registry (export-tables.ts) instead
of three hand-maintained enumerations: each entry carries its backup key,
physical table name, required/optional status and the per-table hooks that
encode real decisions (webhook credential redaction, body_ts stripping,
the shared inline-media budget). Every export validates the registry
against the data DataSource's entity metadata before reading anything: a
registered table the metadata does not know (renamed or dropped entity)
fails the request instead of exporting empty, and a data entity table
with no backup decision fails instead of silently missing every backup.

A parity spec binds the registry to everything that must agree with it:
every data-connection entity table must be exported or explicitly
excluded with a reason, registry tables must exist in entity metadata,
export and import descriptors must cover the same tables in the same
FK-safe order, and the registry keys must equal the published response
DTO keys. Entity files are discovered from the filesystem so a new entity
always forces a decision here.
