+++ src/modules/infra/infra.module.ts
@@ -2,6 +2,7 @@ import { Module, DynamicModule, Type } from '@nestjs/common';
 import { InfraStatusController } from './infra-status.controller';
 import { InfraConfigController } from './infra-config.controller';
 import { InfraDataController } from './infra-data.controller';
+import { InfraDataService } from './infra-data.service';
 import { InfraStorageController } from './infra-storage.controller';
 import { EngineModule } from '../../engine/engine.module';
 import { DockerModule } from '../docker';
@@ -17,9 +18,10 @@ if (process.env.QUEUE_ENABLED === 'true') {
 }
 
 @Module({
-  // SessionModule gives InfraDataController the live-engine registry for the import pre-flight
+  // SessionModule gives InfraDataService the live-engine registry for the import pre-flight
   // orphan check. Its own imports (WebhookModule, StatusStoreModule) never point back here — no cycle.
   imports: [EngineModule, DockerModule, SessionModule, ...queueModules],
   controllers: [InfraStatusController, InfraConfigController, InfraDataController, InfraStorageController],
+  providers: [InfraDataService],
 })
 export class InfraModule {}
